Cleanup and polish the new "PineappleField" API (formerly FastField).
Currently, you can only read the values of PineappleFields,
but eventually you'll be able to set them too.
PineappleFields are faster, safer, and easier to use than the core reflection API.
They should be considered a 'modern' alternative to field objects,
similar to how MethodHandles are to method objects.

Correct the versioning to acurately represent that we're in alpha.
Remove requirement on lombok.